+++ src/OpenInput/Mechanics/Input/IInputTrigger.cs
@@ -1,0 +1,15 @@
+namespace OpenInput.Mechanics.Input
+{
+    public interface IInputTrigger
+    {
+        /// <summary>
+        /// Gets the name associated with this trigger.
+        /// </summary>
+        string Name { get; }
+
+        /// <summary>
+        /// Gets the key that trigger this.
+        /// </summary>
+        InputKey Key { get; }
+    }
+}